+++ Common.Meshing/Constructors/IEdgeMeshConstructor.cs
@@ -11,9 +11,9 @@ namespace Common.Meshing.Constructors
 
         MESH PopMesh();
 
-        void AddVertex(Vector2d pos);
+        void AddVertex(Vector2f pos);
 
-        void AddVertex(Vector3d pos);
+        void AddVertex(Vector3f pos);
 
         void AddEdge(int i0, int i1);
     }
