+++ Common.Meshing/Constructors/ITetrahedronMeshConstructor.cs
@@ -13,7 +13,7 @@ namespace Common.Meshing.Constructors
 
         MESH PopMesh();
 
-        void AddVertex(Vector3d pos);
+        void AddVertex(Vector3f pos);
 
         void AddFace(int i0, int i1, int i2, int i3);
